requirements.txt
google_api_python_client==2.125.0
playwright==1.50.0
pymongo==4.11.3
python-dotenv==1.0.1
pytz==2024.1
